Correcting rotation lambda
The current implementation has a bug where the create function
will never be called resulting in no new credentials being created.
This happend due to the check whether there is already a pending
value which will keep existing after the pending value was promoted
to the current value. Fixed by checking whether the version id of
pending and current are also different.