Improving shared value usage and preparing multiple sdk implementations
Before it was difficult to initialize the shared type for multiple
invocations. Now its possible to use the setup function with access
to the region value.

So far, the rotation lambda only worked with the rusoto sdk. With it
being in maintaince mode, changes are necessary to allow switching
between the old rusoto and the new aws sdk